+++ src/Pages/Platforms/index.tsx
@@ -1,5 +1,6 @@
-import * as React from "react"
+import React from "react"
 import { Route, RouteComponentProps, Switch } from "react-router"
+import { useMeasure } from "react-use"
 import { Col, Row } from "reactstrap"
 
 import { ErddapMap, ErddapPlatformList } from "Features/ERDDAP"
@@ -16,24 +17,25 @@ import { PlatformMatchParams } from "./types"
 /**
  * Top level platform page. Displays region level platform selection if no platform is selected.
  */
-export class PlatformsPage extends React.Component<RouteComponentProps, object> {
-  public render() {
-    const params = urlParams(this.props.location.search)
+export const PlatformsPage: React.FC<RouteComponentProps> = (props: RouteComponentProps) => {
+  const [ref, { height }] = useMeasure<HTMLDivElement>()
+  const params = urlParams(props.location.search)
 
-    const platformId = this.props.match.params.hasOwnProperty("id")
-      ? (this.props.match.params as PlatformMatchParams).id
-      : ""
+  const platformId = props.match.params.hasOwnProperty("id") ? (props.match.params as PlatformMatchParams).id : ""
 
-    let region: Region | undefined
+  let region: Region | undefined
 
-    if (params.region !== undefined) {
-      region = regionList.find((r) => r.slug === params.region)
-    }
+  if (params.region !== undefined) {
+    region = regionList.find((r) => r.slug === params.region)
+  }
 
-    return (
-      <React.Fragment>
-        <Row>
-          <Col sm={{ size: true, order: 6 }}>
+  console.log(height)
+
+  return (
+    <React.Fragment>
+      <Row>
+        <Col sm={{ size: true, order: 6 }}>
+          <div ref={ref}>
             {/* Show list of platforms in a region if no platform is selected */}
             <Switch>
               <Route path={paths.platforms.root} exact={true}>
@@ -44,34 +46,30 @@ export class PlatformsPage extends React.Component<RouteComponentProps, object> 
               </Route>
               <Route path={paths.platforms.platform} component={PlatformInfo} />
             </Switch>
-          </Col>
+          </div>
+        </Col>
 
-          <Col sm={{ size: true, order: 1 }}>
-            <ErddapMap platformId={platformId} boundingBox={region?.bbox} />
-          </Col>
-        </Row>
+        <Col sm={{ size: true, order: 1 }}>
+          <ErddapMap platformId={platformId} boundingBox={region?.bbox} height={region ? "80vh" : height} />
+        </Col>
+      </Row>
 
-        <div style={{ marginTop: "1rem" }}>
-          <Switch>
-            <Route path={paths.platforms.observations}>
-              {(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}
-            </Route>
-            <Route path={paths.platforms.forecast}>
-              {(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}
-            </Route>
-            <Route path={paths.platforms.platform}>
-              {(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}
-            </Route>
-            <Route path={paths.platforms.all}>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
+      <div style={{ marginTop: "1rem" }}>
+        <Switch>
+          <Route path={paths.platforms.observations}>
+            {(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}
+          </Route>
+          <Route path={paths.platforms.forecast}>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
+          <Route path={paths.platforms.platform}>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
+          <Route path={paths.platforms.all}>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
 
-            <Route path={paths.platforms.root} exact={true} component={RootInfo} />
+          <Route path={paths.platforms.root} exact={true} component={RootInfo} />
 
-            <Route>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
-          </Switch>
-        </div>
-      </React.Fragment>
-    )
-  }
+          <Route>{(props) => <PlatformTabs {...(props as PlatformTabsProps)} />}</Route>
+        </Switch>
+      </div>
+    </React.Fragment>
+  )
 }
 
 export default PlatformsPage