+++ CHANGELOG.md
@@ -6,7 +6,13 @@ Additions:
 
 Changes:
 
+- Adjust styles to match up better with the NERACOOS visual guide.
+- Use dashed line to indicate observations in forecast chart.
+- Display color of data in multi-line chart tooltips.
+
 Fixes:
+
+- Keep chart colors from shifting by using a color cycle from NERACOOS visual guide.
 
 ## 0.6.11 - 4/14/2021